share throttle
