+++ front/admin/relocate_users.ts
@@ -1,7 +1,6 @@
 import { concurrentExecutor, removeNulls } from "@dust-tt/types";
-import type { ApiResponse } from "auth0";
 
-import { getAuth0ManagemementClient } from "@app/lib/api/auth0";
+import { getAuth0ManagemementClient, throttleAuth0 } from "@app/lib/api/auth0";
 import { SUPPORTED_REGIONS } from "@app/lib/api/regions/config";
 import { Authenticator } from "@app/lib/auth";
 import { MembershipResource } from "@app/lib/resources/membership_resource";
@@ -33,32 +32,6 @@ makeScript(
 
     const auth = await Authenticator.internalAdminForWorkspace(workspaceId);
     const workspace = auth.getNonNullableWorkspace();
-
-    let remaining = 10;
-    let resetTime = Date.now();
-
-    const throttleAuth0 = async <T>(fn: () => Promise<ApiResponse<T>>) => {
-      if (remaining < rateLimitThreshold) {
-        const now = Date.now();
-        const waitTime = resetTime * 1000 - now;
-        logger.info({ waitTime }, "Waiting");
-        await new Promise((resolve) => setTimeout(resolve, waitTime));
-      }
-
-      const res = await fn();
-      if (res.status !== 200) {
-        logger.error({ res }, "When calling Auth0");
-        process.exit(1);
-      }
-
-      remaining = Number(res.headers.get("x-ratelimit-remaining"));
-      resetTime = Number(res.headers.get("x-ratelimit-reset"));
-
-      const limit = Number(res.headers.get("x-ratelimit-limit"));
-      logger.info({ limit, remaining, resetTime }, "Rate limit");
-
-      return res.data;
-    };
 
     const members = await MembershipResource.getMembershipsForWorkspace({
       workspace,
@@ -95,17 +68,19 @@ makeScript(
         count++;
         logger.info({ user: auth0Id, count }, "Setting region");
         if (execute) {
-          await throttleAuth0(() =>
-            managementClient.users.update(
-              {
-                id: auth0Id,
-              },
-              {
-                app_metadata: {
-                  region: destinationRegion,
+          await throttleAuth0(
+            () =>
+              managementClient.users.update(
+                {
+                  id: auth0Id,
                 },
-              }
-            )
+                {
+                  app_metadata: {
+                    region: destinationRegion,
+                  },
+                }
+              ),
+            { rateLimitThreshold }
           );
         }
       },
